handle import alias
