+++ sample/myawesomemodule1/src/main/res/values/strings.xml
@@ -3,4 +3,5 @@
     <string name="unused_unicode_string">&#169; %1$d xxxxxx v%2$s%3$s</string>
     <string name="used_unicode_string">&#169; ay %1$s &amp;</string>
     <string name="used_full_qualified_import">Fully imported</string>
+    <string name="used_aliased">This is aliased</string>
 </resources>
